Updated to Rust 1.8.0 Syntax
Rust 1.8 has significantly streamlined the use of function pointers
compared to 1.0 nightly. Admittedly this was a year ago. As such, I've
updated this library to use modern function pointers and raw pointers.

Yo Dawg, I heard you like pointers, so I pointer in your pointer so you
can dereference while you point.

Basically converting to a raw pointer for a function pointer results in
a reference to a reference. Thus, you need to dereference the first
reference. It's a bit clunky and that's why there's an unsafe block to
deal with this.

Please note that even though the raw pointer dereferencing happens, this
is *only* to gain the pointer value of the function and use it for
sorting into the binary tree storing handlers. No pointer arithmetic is
done nor any crazy typecasting.